🦄 refactor(Add controller decorator and update routes decorator):
