+++ src/decorators/index.ts
@@ -1,3 +1,4 @@
 export { default as LogInitRequest } from '@decorators/log-request'
 export { default as isConnected } from '@decorators/server-status'
-export { get, controller } from '@decorators/routes'
+export { get, post, patch, del, put } from '@decorators/routes'
+export { controller } from '@decorators/controller'
