+++ src/decorators/routes.ts
@@ -2,10 +2,8 @@ import 'reflect-metadata'
 import express, { Router } from 'express'
 import { PathDir } from '@constants'
 import { conNex } from '@utils'
+import { Method, MetaKey } from '@enum'
 import AppRouter from '@app-router'
-
-// export const router = express.Router()
-const app = express()
 
 /**
  * TODO: implement the linkAuthController function
@@ -17,7 +15,7 @@ export function linkAuthController(target: any) {
 
   for (const key in target.prototype) {
     const routeHandler = target.prototype[key]
-    const path = Reflect.getMetadata('path', target.prototype, key)
+    const path = Reflect.getMetadata(MetaKey.path, target.prototype, key)
 
     if (path) {
       pathRouter.get(path, routeHandler.bind(new target()))
@@ -31,32 +29,23 @@ export function linkAuthController(target: any) {
 
 /**
  *
- * @param routePrefix - The prefix for the route
- * @returns - A function that takes in a target and assigns the routeHandler to the router
- */
-export function controller(routePrefix: string) {
-  return function (target: Function) {
-    const router = AppRouter.instance
-
-    Object.getOwnPropertyNames(target.prototype).forEach((key) => {
-      const routeHandler = target.prototype[key]
-      const path = Reflect.getMetadata('path', target.prototype, key)
-
-      if (path) {
-       let connectedPath = conNex(PathDir.API_ROOT, routePrefix, path)
-        router.get(connectedPath, routeHandler)
-      }
-    })
-  }
-}
-
-/**
- *
- * @param path - The path for the route
+ * @param path  - The path for the route
+ * @method      - GET
  * @returns  - A function that takes in a target and assigns the path to the target
  */
-export function get(path: string) {
-  return function (target: any, key: string, desc: PropertyDescriptor) {
-    Reflect.defineMetadata('path', path, target, key);
+
+export function bindRoute(method: string) {
+  return function (path: string) {
+    return function (target: any, key: string, desc: PropertyDescriptor) {
+      Reflect.defineMetadata(MetaKey.path, path, target, key)
+      Reflect.defineMetadata(MetaKey.method, method, target, key)
+    }
   }
 }
+
+export const get = bindRoute(Method.get)
+export const post = bindRoute(Method.post)
+export const put = bindRoute(Method.put)
+export const del = bindRoute(Method.del)
+export const patch = bindRoute(Method.patch)
+export const options = bindRoute(Method.options)
